+++ ModDebuggingMod/Card.cs
@@ -11,8 +11,8 @@ namespace ModDebuggingMod
         {
             CardInfo info = CardHelper.NewCard(true, "wstl", "wstlcard", "Debug",
                 attack: 1, health: 100, blood: 0, bones: 0, energy: 0, gems: null)
-                .AddAbilities(Ability.ShieldGems)
-                .AddSpecialAbilities(Bind.specialAbility)
+                .AddAbilities(Ability.DeathShield, Ability.DeathShield, Witness.ability)
+                //.AddSpecialAbilities(Bind.specialAbility)
                 //.SetTransformerCardId("Squirrel")//.SetEvolve("Squirrel", 1)
                 .SetPortraits("misterWin_grimora", emissionName: "misterWin_grimora_emission", pixelPortraitName: "buffBell.png")
                 ;