+++ WhistleWind.AbnormalSigils/Abilities/abilityOneSided.cs
@@ -1,5 +1,6 @@
 ﻿using DiskCardGame;
 using InscryptionAPI.Card;
+using InscryptionAPI.Triggers;
 using System.Collections;
 using WhistleWind.AbnormalSigils.Core.Helpers;
 
@@ -19,50 +20,54 @@ namespace WhistleWind.AbnormalSigils
                 modular: true, opponent: true, canStack: true).Id;
         }
     }
-    public class OneSided : AbilityBehaviour
+    public class OneSided : AbilityBehaviour, IModifyDamageTaken
     {
         public static Ability ability;
         public override Ability Ability => ability;
 
-        public override bool RespondsToDealDamage(int amount, PlayableCard target)
-        {
-            if (amount > 0 && target != null && !target.Dead)
-                return CheckValid(target);
+        private bool activate = false;
 
-            return false;
+        public override bool RespondsToDealDamage(int amount, PlayableCard target) => activate;
+        public bool RespondsToModifyDamageTaken(PlayableCard target, int damage, PlayableCard attacker, int originalDamage)
+        {
+            return base.Card == attacker && CheckValid(target);
         }
 
         public override IEnumerator OnDealDamage(int amount, PlayableCard target)
         {
+            activate = false;
             yield return base.PreSuccessfulTriggerSequence();
             yield return base.LearnAbility(0.4f);
         }
+        public int OnModifyDamageTaken(PlayableCard target, int damage, PlayableCard attacker, int originalDamage)
+        {
+            activate = true;
+            damage++;
+            return damage;
+        }
+
         private bool CheckValid(PlayableCard target)
         {
             // if target has no Power, if this card can submerge or is facedown (cannot be hit), return true by default
-            if (target.Attack == 0 || base.Card.HasAnyOfAbilities(Ability.Submerge, Ability.SubmergeSquid) || base.Card.FaceDown)
+            if (target.Attack == 0)
                 return true;
 
-            // if this card doesn't have Sniper or Marksman (will attack opposing)
-            if (base.Card.LacksAbility(Ability.Sniper))
-            {
-                // if this card has Bi or Tri Strike, check whether the opponent has it too
-                if (base.Card.HasAbility(Ability.SplitStrike) || base.Card.HasTriStrike())
-                    return !(target.HasAbility(Ability.SplitStrike) || target.HasTriStrike());
+            if (target.HasAbility(Ability.Flying) && base.Card.LacksAbility(Ability.Reach))
+                return true;
 
-                // otherwise, return whether the opponent can attack this card (won't attack directly or is blocked)
-                return target.CanAttackDirectly(base.Card.Slot) || target.AttackIsBlocked(base.Card.Slot);
-            }
-            // if the target is opposing this card
-            if (target.Slot == base.Card.Slot.opposingSlot)
-                return target.CanAttackDirectly(base.Card.Slot) || target.AttackIsBlocked(base.Card.Slot);
+            // Persistent check
+            if ((base.Card.HasAbility(Ability.TailOnHit) && !base.Card.Status.lostTail) ||
+                base.Card.HasAbility(Ability.PreventAttack) ||
+                base.Card.HasAnyOfAbilities(Ability.Submerge, Ability.SubmergeSquid) || base.Card.FaceDown)
+                return target.LacksAbility(Persistent.ability);
 
-            // if the target is in an opposing adjacent slot
-            if (Singleton<BoardManager>.Instance.GetAdjacentSlots(base.Card.Slot.opposingSlot).Contains(target.Slot))
-                return target.LacksAbility(Ability.SplitStrike) || !target.HasTriStrike();
+            // Shield checks
+            if (base.Card.HasShield())
+                return target.LacksAllAbilities(Piercing.ability, Ability.Sharp, Reflector.ability);
 
-            // otherwise return true
-            return true;
+            return !target.GetOpposingSlots().Contains(base.Card.Slot);
         }
+
+        public int TriggerPriority(PlayableCard target, int damage, PlayableCard attacker) => 0;
     }
 }