+++ WhistleWind.AbnormalSigils/Abilities/abilityWitness.cs
@@ -5,7 +5,7 @@ using System.Collections;
 using System.Linq;
 using UnityEngine;
 using WhistleWind.AbnormalSigils.Core.Helpers;
-
+using WhistleWind.AbnormalSigils.StatusEffects;
 using WhistleWind.Core.AbilityClasses;
 
 namespace WhistleWind.AbnormalSigils
@@ -17,7 +17,7 @@ namespace WhistleWind.AbnormalSigils
             const string rulebookName = "Witness";
             const string rulebookDescription = "Pay 2 Bones to increase the selected creature's Health by 2 and their taken damage by 1. This effect stacks up to 3 times.";
             const string dialogue = "The truth will set you free.";
-            const string triggerText = "The chosen creature beholds [creature] and is reborn.";
+            const string triggerText = "The chosen beholds [creature] and is reborn.";
             Witness.ability = AbnormalAbilityHelper.CreateActivatedAbility<Witness>(
                 "sigilWitness",
                 rulebookName, rulebookDescription, dialogue, triggerText, powerLevel: 1).Id;
@@ -32,37 +32,41 @@ namespace WhistleWind.AbnormalSigils
         public override int StartingBonesCost => 2;
         public override bool IsInvalidTarget(CardSlot slot)
         {
-            // slot is occupied and isn't an opposing slot
-            if (slot.Card != null && slot.IsPlayerSlot != base.Card.OpponentCard)
+            if (base.IsInvalidTarget(slot))
+                return true;
+
+            if (slot.Card.OpponentCard == base.Card.OpponentCard)
             {
-                int prudence = slot.Card.Info.GetExtendedPropertyAsInt("wstl:Prudence") ?? 0;
-                return prudence >= 3;
+                var component = slot.Card.GetStatusEffect<Prudence>();
+                if (component != null)
+                    return component.EffectSeverity > 3;
+
+                return false;
             }
-            return base.IsInvalidTarget(slot);
+            return true;
         }
 
         public override bool CanActivate()
         {
-            // Are there other cards to affect, are any of them valid targets, and is this card on the player's side
-            foreach (var slot in BoardManager.Instance.GetSlotsCopy(!base.Card.OpponentCard).Where((CardSlot slot) => slot.Card != base.Card))
+            // If there's a valid target on the owner's side
+            foreach (var slot in BoardManager.Instance.GetSlotsCopy(!base.Card.OpponentCard))
             {
-                if (slot.Card != null)
-                {
-                    // if there is at least one other card that can take more Prudence
-                    int prudence = !(slot.Card.Info.GetExtendedPropertyAsInt("wstl:Prudence") != null) ? 0 : (int)slot.Card.Info.GetExtendedPropertyAsInt("wstl:Prudence");
-                    if (prudence < 3)
-                        return true;
-                }
+                if (!IsInvalidTarget(slot))
+                    return true;
             }
             return false;
         }
 
         public override IEnumerator OnValidTargetSelected(CardSlot slot)
         {
-            int prudence = !(slot.Card.Info.GetExtendedPropertyAsInt("wstl:Prudence") != null) ? 0 : (int)slot.Card.Info.GetExtendedPropertyAsInt("wstl:Prudence");
-            yield return slot.Card.Info.SetExtendedProperty("wstl:Prudence", prudence + 1);
-            slot.Card.HealDamage(2);
             slot.Card.Anim.StrongNegationEffect();
+            slot.Card.HealDamage(2);
+
+            if (slot.Card.HasStatusEffect<Prudence>())
+                slot.Card.UpdateStatusEffectCount<Prudence>(1, false);
+            else
+                slot.Card.AddStatusEffectToCard<Prudence>();
+
             yield return new WaitForSeconds(0.4f);
             yield return base.LearnAbility();
         }